Add initial implementation of CompletionRequests
Adds a builder (CompletionRequest) for building the request
to be sent to the OpenAI API and return the generated response
(CompletionResponse).

All request options found [here](https://platform.openai.com/docs/api-reference/completions/create)
are supported except for streaming the request (this will be added in the future)

This commit also adds tests that require setting the OPENAI_API_KEY env
variable. The tests do submit real requests to the OpenAI API and will consume
tokens. However, the tests are written againts the cheapest (and most simple)
model ada.

Signed-off-by: Andrea Gunderson <agunde@embyr.ai>